refactor(Columns): position 타입 Long -> double 전환

+++ src/main/java/com/project/skilled_project/domain/columns/dto/request/ColumnsCreateRequestDto.java
@@ -7,6 +7,6 @@ public class ColumnsCreateRequestDto {
 
   private Long boardId;
   private String title;
-  private Long position;
+  private double position;
 
 }
